Code corrections driven by database definition changes
git-svn-id: https://svn.mothershipdev.com:9443/svn/PaymentSense_MothershipV2_Core/branches/DiallerService@275 0f896bb5-3ab0-2d4e-82a7-85ef10020915

+++ PS.Mothership.Core/PS.Mothership.Core.Common.Rellaid/Dto/DiallerCallDto.cs
@@ -48,7 +48,7 @@ namespace PS.Mothership.Core.Common.Rellaid.Dto
         [DataMember]
         public Guid? ConsultOriginSipCallGuid { get; set; }
         [DataMember]
-        public int? RecorderCallId { get; set; }
+        public long? RecorderCallId { get; set; }
         [DataMember]
         public Guid? ProspectingCampaignResponseTapCallGuid { get; set; }
         [DataMember]
